requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
peewee==3.17.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.1

